Split out objective function so we can also call for GLOP
